requirements.txt
numpy
matplotlib
PyYAML
pandas
scipy
ipython
seaborn
matplotlib-inline
wand
termcolor
typing
icecream
executing==2.0.0
